streamlitanalytics-master/requirements.txt
streamlit>=1.31.0
databricks-sql-connector>=3.0.0
pandas>=2.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
